nanobot-orchestrator/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
redis>=5.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
nanobot-ai>=0.1.5
